+++ .gitignore
@@ -1,1 +1,11 @@
 /.idea
+target/
+.idea/
+test/
+*.iml
+.DS_Store
+src/.DS_Store
+src/main/.DS_Store
+*.class
+*.jar
+*.war
