+++ pom.xml
@@ -47,7 +47,12 @@
       <artifactId>jackson-dataformat-xml</artifactId>
       <version>2.10.0</version>
     </dependency>
-
+    <dependency>
+      <groupId>log4j</groupId>
+      <artifactId>log4j</artifactId>
+      <version>1.2.17</version>
+      <scope>compile</scope>
+    </dependency>
     <dependency>
       <groupId>org.apache.commons</groupId>
       <artifactId>commons-lang3</artifactId>
