+++ src/main/resources/config.properties
@@ -1,0 +1,3 @@
+pgHost = localhost
+pgLogin = postgres
+pgPassword =
